Relayer verifies refund before propagating transaction

+++ universal-login-relayer/lib/services/IdentityService.js
@@ -1,17 +1,18 @@
 import Identity from 'universal-login-contracts/build/Identity';
-import {addressToBytes32} from '../utils/utils';
+import {addressToBytes32, isEnoughGasLimit, hasEnoughToken} from '../utils/utils';
 import ethers, {utils, Interface} from 'ethers';
 import defaultDeployOptions from '../config/defaultDeployOptions';
 
 
 class IdentityService {
-  constructor(wallet, ensService, authorisationService, hooks) {
+  constructor(wallet, ensService, authorisationService, hooks, provider) {
     this.wallet = wallet;
     this.abi = Identity.interface;
     this.ensService = ensService;
     this.authorisationService = authorisationService;
     this.codec = new utils.AbiCoder();
     this.hooks = hooks;
+    this.provider = provider;
   }
 
   async create(managementKey, ensName, overrideOptions = {}) {
@@ -31,17 +32,26 @@ class IdentityService {
   }
 
   async executeSigned(contractAddress, message) {
-    const contract = new ethers.Contract(contractAddress, this.abi, this.wallet);
-    const addKeySighash = new Interface(Identity.interface).functions.addKey.sighash;
-    if (message.to === contractAddress && message.data.slice(0, addKeySighash.length) === addKeySighash) {
-      const [address] = (this.codec.decode(['bytes32', 'uint256', 'uint256'], message.data.replace(addKeySighash.slice(2), '')));
-      const key = utils.hexlify(utils.stripZeros(address));
-      await this.authorisationService.removeRequest(contractAddress, key);
-      const transaction = await contract.executeSigned(message.to, message.value, message.data, message.nonce, message.gasToken, message.gasPrice, message.gasLimit, message.signature);
-      this.hooks.emit('added', key);
-      return transaction;
+    const {data} = new Interface(Identity.interface).functions.executeSigned(message.to, message.value, message.data, message.nonce, message.gasToken, message.gasPrice, message.gasLimit, message.signature);
+    const transaction = {
+      value: 0,
+      to: contractAddress,
+      data,
+      ...defaultDeployOptions
+    };
+    const estimateGas = await this.wallet.estimateGas(transaction);
+    if (isEnoughGasLimit(estimateGas, message.gasLimit) && await hasEnoughToken(message.gasToken, contractAddress, message.gasLimit, this.provider)) {
+      const addKeySighash = new Interface(Identity.interface).functions.addKey.sighash;
+      if (message.to === contractAddress && message.data.slice(0, addKeySighash.length) === addKeySighash) {
+        const [address] = (this.codec.decode(['bytes32', 'uint256', 'uint256'], message.data.replace(addKeySighash.slice(2), '')));
+        const key = utils.hexlify(utils.stripZeros(address));
+        await this.authorisationService.removeRequest(contractAddress, key);
+        const sentTransaction = await this.wallet.sendTransaction(transaction);
+        this.hooks.emit('added', key);
+        return sentTransaction;
+      }
+      return await this.wallet.sendTransaction(transaction);
     }
-    return await contract.executeSigned(message.to, message.value, message.data, message.nonce, message.gasToken, message.gasPrice, message.gasLimit, message.signature);
   }
 }
